refactor: move run_validation_operator
